Major concurrency fix, remove .git
While modifying the repository.js `fetch` method to remove .git directory from commitish, I noticed that one of the functional tests was breaking. This lead me to find a major issue where if multiple requests were being made to the same repo/commitish while it was being fetched, then the subsequent requests would fail. This fixes the problem by rearranging when and how we spin-lock the cache directories, and adding additional logic to see if we're already cached after a lock is acquired.